Editor Additions (#479)
* favoriting and pick block

favoriting needs to not break after previewing (maybe make it persistent in the settings file?)
pick block needs to pick the icon, too. and preferably scroll to it as well

* final star anim + automatic icon game switching

before i make any more changes im making a checkpoint here cuz i know it works
* i want to add a way to specify which event SwitchGame() will switch to (because that's a cool feature for pick block)
* i'll have to figure out how to auto scroll to the game when the icon is selected
* the star now fully works, even between preview switches :+1:

* fix the rest of the stuff

the event name gets colored correctly and hidden games are skipped over, but still loaded.
also i built mm ass buns

* tweaks + zoom and sorting

this stuff will be in the pr desc so it doesn't matter

* oop one more thing

* icons look better now :D

mipmaps to the rescue

* double date fix

+++ Assets/Scripts/Games/PajamaParty/PajamaParty.cs
@@ -101,7 +101,7 @@ namespace HeavenStudio.Games
         static double WantFiveJump = double.MinValue;
         static double WantThrowSequence = double.MinValue;
         static double WantSleepSequence = double.MinValue;
-        static double WantInstantSleep = Single.MinValue;
+        static double WantInstantSleep = double.MinValue;
         static bool WantSleepType = false;
         static int WantSleepAction = (int) PajamaParty.SleepType.Normal;
         static int WantInstantSleepAction = (int) PajamaParty.SleepType.Normal;
